Makes the ADIRS align timer respect the in-game time.
The ADIRS timer was using real world time to update its timer, which
means it completely disregarded sim acceleration. This patch makes
the timer take acceleration in account. In essence, the timer now
follows the same time as the aircraft clock.

There are other places that use similar real world timers instead
of in-game time and we can change those as well in subsequent PRs
if desired.