Update A320_Neo_CDU_MainDisplay.js
This makes the cruise a tiny bit faster (according to the cost index) and the managed descent speed is not ~240kts from top down anymore, but around 285-290kts (speed limit below 10.000 still kept). I didn't touch the CLB speed here, cause - for some reason - it doesn't do anything (even if there are numbers for CLB around this section).This reduces the climb speed from ~320 to ~290kts. I'm not sure about this number. I did it that way, because the Asobo AP seems to handle mach speeds incorrectly. Even if the speed bug is switched to mach, the AP seems to still hold KIAS. So the 290kts here are mainly to smooth the transition between CLB and CRZ. Suggestions of other / more realistic values are appreciated for the future (when AP is working correctly).