Update Deice.xml
I was using this file on my TBM 930 project, and I only edited a single value to decrease the Inertial Separator effect on TRQ from 25% to 22%. Since you guys added this to the project, it's started conflicting with mine and I decided to remove it.
I think this value don't make any difference to the A32NX project, since the Airbus doesn't have an Inertial Separator system. What do you guys think?