fix: braking while taxiing should not turn autobrake off (#1762)
* Fix for bug 255: Braking while taxiing turn autobrake off.
https://github.com/flybywiresim/a32nx/issues/255

Changes the autobrake push button arming behavior so that braking while the
autobrakes are not engaged will not disarm them. Additionally, autobrakes will
automatically disarm upon landing gear retraction and once the autobrakes have
disengaged after engaging. Mostly this is correct, except that there should
be a 10 second delay after landing gear retraction before the autobrakes disarm
so I will be looking into implementating that.

* Added a delay before disarming the autobrake after the landing gear is
retracted. Trying to approximate a delay of 10 seconds.

Disabled setting autobrake max in the air

Added a check to re-enable disarming the autobrakes when the landing gear
retracts if the aircraft touches down.

Added code to ensure that the auto brakes can be set correctly in response
to the .flt files.

* Added comments and a timer to ensure that we are on the ground for
approximately 10 seconds before seeking to disarm the autobrake switches. The
latter is currently necessary because the autobrakes engage immediately upon
touchdown, then disengage if the aircraft bounces, then reengage. We don't want
to disarm the switch due to a bounce.

* Removed an unneeded call to set L:A32NX_DISARM_AUTOBRAKES_ON_LG_RETRACT=1

* Remove my name from the template comments as it is unneeded.

* Add changes for pull request.

* Ensure that if autobrakes are set in the air, trying to push MAX
causes the original setting to be retained.

* Revert "Ensure that if autobrakes are set in the air, trying to push MAX"

* Pressing MAX autobrake button while the aircraft has no effect. Previously
pressing MAX would disarm autobrakes if they were set to an allowable value.
Now if autobrakes are set to LO or MED, those settings will be preserved if
MAX is pressed.

Co-authored-by: Geoff Darst <geoffda@hotmail.com>
Co-authored-by: Benjamin Dupont <benjozorkfr@gmail.com>