Optimize performance by sequencing updates (+ ~50% fps) (#2976)
* Optimize performance by sequencing updates

Instead of updating everything in the same frame, this introduces
UpdateThrottler that aims to throttle updates depending on the context.

Example: While the Speed indicator needs to very snappy, other things
like information about the electric system can update at a much reduced
rate.

This is a little bit like the in-game option to change the glass cockpit
refresh rate, however:
- With this patch, we do it individually, such that we have different
refresh rates as mentioned above.
- Furthermore, this tries to space out updates from different elements
among different frames. With the in-game setting, everything updates in
one frame, but then nothing updates in the next frame. With this patch,
we try to space out upates for different instrument among all frames,
leading to a lot less spikey frame times. This leads to a much better
experience in VR and VRR (variable refresh rate (G-Sync or FreeSync))
capable monitors.

It's recommended to set in-game glass cockpit refresh rate to high for
this to work best, but it also works for lower refresh rates.

* Apply suggestions from code review

Co-authored-by: Benedict Etzel <developer@beheh.de>

* Apply suggestions from code review

Co-authored-by: Benedict Etzel <developer@beheh.de>

* Fix self test flickering

* Apply suggestions from code review

Co-authored-by: DerL30N <leon.beeser@yahoo.de>

* Remove throttling for SAI pressure indicator

Co-authored-by: pepperoni505 <49950581+pepperoni505@users.noreply.github.com>
Co-authored-by: Benedict Etzel <developer@beheh.de>
Co-authored-by: DerL30N <leon.beeser@yahoo.de>
Co-authored-by: Benjamin Dupont <benjozorkfr@gmail.com>