fix: APU SWITCH temporarily became false
Previously I wrote that something outside of our code is setting the APU
SWITCH to 0. It turns out this is related to the FUEL_SYSTEM_VALVE
leading to the APU. When it is closed and the APU is running, MSFS will
turn the sim's APU off. By also opening the fuel valve at the same time
as starting the sim's APU we fix problems related to this "flickering".