feat: Decals better fonts and improved emissive (#3693)
* Decals font fix

Put the proper font on the cockpit decals and remade the airframe decals.

* Update CHANGELOG.md

* Update A320NEO_COCKPIT_DECALSTEXT_EMIS.PNG.dds

Lightened the blue and made the MAN white.

* Update A320NEO_COCKPIT_DECALSTEXT_ALBD.TIF.dds

Squib was not there. This has been rectified.

* Made ALTN white.

* Update A320NEO_COCKPIT_DECALSTEXT_ALBD.TIF.dds

Added missing ()

* Update A320NEO_COCKPIT_DECALSTEXT_EMIS.PNG.dds

Made the blue dots white.

* Decals font fix

Put the proper font on the cockpit decals and remade the airframe decals.

* Update CHANGELOG.md

* Update A320NEO_COCKPIT_DECALSTEXT_EMIS.PNG.dds

Lightened the blue and made the MAN white.

* Update A320NEO_COCKPIT_DECALSTEXT_ALBD.TIF.dds

Squib was not there. This has been rectified.

* Made ALTN white.

* Update A320NEO_COCKPIT_DECALSTEXT_ALBD.TIF.dds

Added missing ()

* Update A320NEO_COCKPIT_DECALSTEXT_EMIS.PNG.dds

Made the blue dots white.

* Update A320NEO_COCKPIT_DECALSTEXT_ALBD.TIF.dds

Co-authored-by: pepperoni505 <49950581+pepperoni505@users.noreply.github.com>
Co-authored-by: Will Pine <wpine215@gmail.com>