fix(elec): lack of power not detected in same tick (#4076)
The frequency and potential of the power sources are only known at the
end of a tick, due to them being directly related to the power
consumption (large changes can cause spikes and dips). However, the
decision if a power source can supply power is made much earlier in the
tick. This is especially of great consequence when the power source no
longer supplies potential but the previous tick's frequency and
potential are still normal.

To fix this, I include whether or not the power source
should_provide_output in the output_within_normal_parameters function. 
should_provide_output is always up-to-date (doesn't have a one tick delay).

As a consequence, a power source losing power will be detected
immediately, but a power source just starting providing output will only
be within normal conditions after a single tick. I deem this acceptable.