[BUG] FOB information is reset when Refuel *LOAD pressed in MCDU mid-flight #3502
Aircraft can be fueled via ATSU only when ground control is ON.
The reason for this idea is that ground control is only ON when aircraft is on ground. When control is OFF fueling via ATSU is not working. Ground control is not activate while taxiing, flying etc..

Here comes another issue:
Right now ground control can be activated in flight. Someone needs to work on this.

If there is any problem regarding my solution please get in contact with me.

Ersin