fix(nd): alt constraint circle
Alt constrained waypoints are displayed with a circle around them. The
circle should be white when the constraint is not taken into account/NAV
is not engaged, magenta when it will be met, and amber when it will be
missed. The met/missed predictions are not available yet (VNAV), so for
now we just display them in white.