fix: workaround for #5487 in regards to altitude issues
This commit introduces a workaround for the altimeter issues in MSFS.

The issue in MSFS is that the pressure altitude and indicated altitude
are not correctly calculated. Additionally when the barometer setting
is set to STD both altitudes are not equal.

The impact is that when sea level pressure changes the error between
pressure and indicated altitude varies. The autopilot is using the
pressure altitude as reference when in ALT HOLD mode. Due to this,
when flying through different pressure zones the AP will divert from the
altitude it should follow.

The changes that are made with this commit are as follows:
- add an additional altimeter (index 2)
- force this altimeter to baro setting STD
- use the indicated altitude as pressure altitude

This fixes the autopilot to divert from the altitude when in ALT HOLD.

Known issues that might remain:
- pressure altitude is still wrong compared to ICAO standard atmosphere
- ATC / VATSIM not accepting the altitude as correct flight level