fix: refactor NXDataStore
usePersistentPropertyWithDefault was not needed (usePersistent property could already do what it did).
There was a bug in usePersistentProperty which stopped it from ever setting the default in the datastore... and it is bound to be racy if you rely on it across instruments etc. anyway... removed.

provide userPersistentNumberProperty for convenience.

Added overloads to avoid possibly undefined issues for callers when they provide a default. Required eslintrc rules to be changed as they were not typescript aware.