Fix reverse thrust toggle
This small fix changes the behaviour of the reverse thrust toggle.
When a the reverse is either hold or toggle, the reverse thrust lever goes all
the way down (-20). This fix support controllers that uses this action with a
toggle button or holding it, instead of flipping the thrust axis.

Signed-off-by: Erick Cardona <erick0zcr@gmail.com>