refactor(ND): stateful mapParams
mapParams in PlanMode are now fully statefull. A new MapParameters
is created for new computation, which kicks in re-render.

The (general) problem of mapParams is that only one instance was kept
and recomputed, which needed children to re-render on different prop changes.

Note that MapParameters objects are relatively small and should not impact GC.