fix: clean up adr altitude labels and remove isa deviation
I have not implemented the baro corrected alt labels correctly yet as they still rely on the sim altimeters => they change to pressure alt when STD is selected, but at least the proper labels are there so other subsystems canhook up to the proper labels and end up correct in future.

There are things that probably ought to be using pressure alt rather than baro corrected alt, but I didn't go into researching all of that. That's a result of having the ADR labels messed up since the beginning.