Fixing Remote SimBridge IP not resetting after selecting 'My PC'
Added extra logic to the getSimBridgeUrl() function to default to 'localhost' for the IP address when the SimBridge is in local mode rather than remote. This way, a user switching between 'This PC' and 'Remote PC' under the FlyPad Sim Options doesn't need to manually change the IP to 'localhost' for local SimBridge to work and can retain the saved IP for switching back.