feature: panel engraving font for texture authoring
Recreation of the Airbus 320 series panel font as found on most engraved panels. Made from manually tracing the characters of the decals sheet and cross checking various photo and video references.

The lower-case characters and several of the symbols are made up to match the theme and what an engraving tool could achieve since I have no references for them, but I wanted to make the font somewhat more generally useful as well.