Add files via upload
This is my first time in Typescript and JavaScript so bare with me. Feel free to help me out and add me with an @ when changing the code. I wish to understand these two languages but Im just at the start.

I tried to edit the columns first so there is **NO** proof so far that this even worked out