feat(a380x/pfd): velocity vector (#8966)
* feat(a380x/pfd): velocity vector

* fix: Arinc429ConsumerSubject

The typing on Arinc429MutateFunc was incorrect, as it actually takes an Arinc429WordData as arg b, and that does not have a word member. With an Arinc429Register consumer it would work anyway. It accepted an Arinc429Word consumer though (because it implements Arinc429WordData), which silently failed to work, setting the internal register to undefined.

Technically the typings are still wrong, as arg a of the mutate function is also Arinc429WordData which does not implement a set method, but practically the initialValue is fixed as an Arinc429Register which does.

This whole mess could be cleaned up by getting rid of Arinc429Word so I have deprecated it here. It's whole design is very bad for CoherentGT as it requires constant allocations.

* fix(a380x/pfd): fpv/vv layer order

* fix: start out paused