fix(a380x): Baro Knob Fix (#9252)
* Update A32NX_Interior_Misc.xml

Changed pull and push functionality for baro knob. Pulling the knob now toggles baro mode and pushing the knob sets it to STD

* Update A32NX_Interior_FCU.xml

Changed the functionality of the baro knob. Pushing the knob now sets it to STD mode and pulling the knob sets it to baro mode.

* Update A32NX_Interior_FCU.xml

* Update A32NX_Interior_FCU.xml

Changed the tooltips content for the pushing/pulling of the baro knob.

* Update A32NX_Interior_Misc.xml

Removed unused code for pushing/pulling the baro knob.

* Update CHANGELOG.md

* Update CHANGELOG.md

---------